+++ src/CleanRegex/Exception/PlaceholderFigureException.php
@@ -1,6 +1,6 @@
 <?php
 namespace TRegx\CleanRegex\Exception;
 
-class PlaceholderFigureException extends \RuntimeException
+class PlaceholderFigureException extends \RuntimeException implements PatternException
 {
 }