+++ ChangeLog.md
@@ -7,6 +7,7 @@ Incoming
 * Breaking changes
     * Removed `SubjectNotMatchedException.getSubject()`.
     * Moved `ChainedReplace` to `TRegx\CleanRegex\Replace\ `
+    * Moved `PatternTemplate` to `TRegx\CleanRegex\ `
 * Other
     * Exception `PlaceholderFigureException` now extend correctly implements `PatternException`.